% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata-api.R
\name{fetch_catalog}
\alias{fetch_catalog}
\title{Fetch the available catalog from healthdata.gov}
\usage{
fetch_catalog(agency = NULL, keyword = NULL)
}
\arguments{
\item{agency}{name of agency to pull in API call.
This requires the full agency name in proper case.
Function list_agencies() can be used for a list of
available agencies. Defaults to NULL.}

\item{keyword}{keyword(s) to pull in API call.
This argument can be supplied on its own or with an agency argument.
The keyword argument uses regular expression to detect presence of
pattern in a string. Function get_keywords() can be used for a list
of exact keywords available in API. Defaults to NULL.}
}
\value{
a \code{\link[tibble:tibble]{tibble}} with descriptive metadata of
the available catalog including a nested list-column of the
data distribution.
}
\description{
This function allows users to pull the catalog in its entirety,
filter by agency, keyword(s), or both
}
\examples{
\dontrun{
fetch_catalog("Centers for Disease Control and Prevention")

cdc_alc <- fetch_catalog("Centers for Disease Control and Prevention",
  keyword = "alcohol|alcohol use"
)
}
}
