% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_api.R
\name{fetch_csv}
\alias{fetch_csv}
\title{Function to fetch data from the healthdata.gov API
This function will pull data if available from healthdata.gov in csv format}
\usage{
fetch_csv(catalog)
}
\arguments{
\item{catalog}{this argument requires a catalog (tibble) created with fetch_catalog()}
}
\value{
a nested tibble with descriptive metadata of a data product and a list-column
        of the available dataset.
}
\description{
Function to fetch data from the healthdata.gov API
This function will pull data if available from healthdata.gov in csv format
}
\examples{
\dontrun{
cdc_alc <- fetch_catalog("Centers for Disease Control and Prevention",
                          keyword = "alcohol|alcohol use")

nested_df_alc <- cdc_alc \%>\%
   slice(1:2) \%>\%  # pull Alzheimer's and Chronic Indicators datasets
   fetch_csv()
}
}
