% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_agencies}
\alias{get_agencies}
\title{Extract names of agencies cataloged in the healthdata.gov API}
\usage{
get_agencies(agency = NULL)
}
\arguments{
\item{agency}{enter partial name or initials of agency to search
catalog for string pattern.
Defaults to \code{NULL} and pulls all agencies cataloged in API}
}
\value{
a \code{\link[tibble:tibble]{tibble}} of names
}
\description{
This function will download a list of Agencies listed as having
an available data product in the
\href{https://healthdata.gov/}{healthdata.gov} data API endpoint.
}
\examples{
\dontrun{
get_agencies()

get_agencies("institute")

get_agencies("substance abuse")
}
}
