% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_keywords}
\alias{get_keywords}
\title{Helper function to extract keywords listed in API}
\usage{
get_keywords(agency = NULL)
}
\arguments{
\item{agency}{enter full agency name in title case to pull keywords tagged in agency's products.
Defaults to NULL and pulls all keywords cataloged}
}
\value{
a tibble with publisher (agency) name(s) and respective keywords
}
\description{
Helper function to extract keywords listed in API
}
\examples{
\dontrun{
get_keywords()
get_keywords("Centers for Disease Control and Prevention")
}
}
