% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdatacsv-package.R
\name{healthdatacsv}
\alias{healthdatacsv}
\title{Access and download data from the healthdata.gov catalog}
\description{
Access and download data from the healthdata.gov catalog
}
\author{
Ivan Castro
}
