% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_agencies}
\alias{list_agencies}
\title{Function to extract names of agencies cataloged in the healthdata.gov API}
\usage{
list_agencies(agency = NULL)
}
\arguments{
\item{agency}{enter partial name or initials of agency to search
catalog for string pattern.
Defaults to NULL and pulls all agencies cataloged in API}
}
\value{
a [tibble][tibble::tibble-package] of names
}
\description{
Function to extract names of agencies cataloged in the healthdata.gov API
}
\examples{
\dontrun{
list_agencies()
list_agencies("institute")
list_agencies("substance abuse")
}
}
