% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata-api.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from the healthdata.gov API}
\usage{
fetch_data(catalog)
}
\arguments{
\item{catalog}{this argument requires a catalog (tibble) created
with \code{fetch_catalog()}}
}
\value{
a \code{\link[tibble:tibble]{tibble}} with descriptive metadata
of a data product and a list-column of the available dataset.
}
\description{
This function will download data from
\href{https://healthdata.gov/}{healthdata.gov},
if available in csv format
}
\examples{
\dontrun{
cdc_alc <- fetch_catalog("Centers for Disease Control and Prevention",
  keyword = "alcohol|alcohol use"
)

nested_df_alc <- cdc_alc \%>\%
  dplyr::slice(1:2) \%>\% # pull Alzheimer's and Chronic Indicators datasets
  fetch_data()
}
}
