% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_agencies}
\alias{get_agencies}
\title{Extract names of agencies cataloged in the healthdata.gov API}
\usage{
get_agencies(agency = NULL)
}
\arguments{
\item{agency}{enter partial name or initials of agency to search
catalog for string pattern.
Defaults to NULL and pulls all agencies cataloged in API}
}
\value{
a \code{\link[tibble:tibble]{tibble}} of names
}
\description{
Extract names of agencies cataloged in the healthdata.gov API
}
\examples{
\dontrun{
get_agencies()

get_agencies("institute")

get_agencies("substance abuse")
}
}
