% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_agencies}
\alias{list_agencies}
\title{Helper function  to extract names of agencies cataloged in the healthdata.gov API}
\usage{
list_agencies(namecheck = NULL)
}
\arguments{
\item{namecheck}{enter partial name or initials of agency to search
catalog for string pattern. Use title case.
Defaults to NULL and pulls all agencies cataloged in API}
}
\value{
a [tibble][tibble::tibble-package] of names
}
\description{
Helper function  to extract names of agencies cataloged in the healthdata.gov API
}
\examples{
\dontrun{
list_agencies()
list_agencies("Institute")
list_agencies("Substance Abuse")
}
}
